Fix CI failure: overwrite join.rs to ensure clean state
CI was failing with spurious 'unused import' and 'unused variable' errors likely due to a merge conflict or state inconsistency. This commit overwrites `relvar-core/src/algebra/join.rs` with the correct, verified content to ensure consistency.
